+++ components/Card.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image"
-import Link from "@/components/Link"
+import Link from "@/components/CustomLink"
 
 const Card = ({
   title,
