+++ components/CategoryCarousel.tsx
@@ -1,4 +1,4 @@
-import Link from "@/components/Link"
+import Link from "@/components/CustomLink"
 import ImageFixed from "next/image"
 import Image from "@/components/CustomImage"
